new Frame stuff so Frames can alloc and write directly to their own Texture buffers, eliminating a copy of frame data.  some code cleanup. switching to 32-bit vars from 16-bit in a few places.

+++ src/devices/displaypp/VideoScannerII.hpp
@@ -93,7 +93,7 @@ protected:
     uint8_t *ram;
 
     ScanBuffer *frame_scan = nullptr;
-    uint16_t scan_index = 0;
+    uint32_t scan_index = 0;
 
     // floating bus video data
     uint8_t   video_byte;
